[Release] 2025-08-14 배포

See merge request s13-webmobile1-sub1/S13P11A404!258

+++ frontend/BE/src/main/java/com/a404/duckonback/repository/RoomRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     List<Room> findByCreator_Id(Long id);       // 해당 유저가 만든 방
-    List<Room> findByArtist_ArtistId(Integer artistId); // 특정 아티스트 방
+    List<Room> findByArtist_ArtistId(Long artistId); // 특정 아티스트 방
 }
